build: cut the warning volume from 269 to 0, and make format errors fatal
The balance-GA bug in a4ea4dc was reported by clang on every single build and
nobody read it, because a whole-tree build emitted 269 warning lines. This
makes the build quiet enough that the next one is visible.

Where the 269 came from, and what happened to them:

  116  doctest's <ciso646>, one per test TU. It includes it under clang purely
       to sniff for _LIBCPP_VERSION; libstdc++ 16 answers with a #warning that
       the header is not standard since C++20. Silenced per test target --
       clang shows an explicit #warning even from a SYSTEM header, and no
       project code uses #warning.
   59  stb_image_write.h. Vendored headers are now included as SYSTEM, so
       third_party stops being warned about as if it were ours.
   44  Application.cpp, all int32_t loop counters subscripting
       std::array<..., 256>. The counters are size_t now.
   15  GreatPeople.cpp, rows mixing positional and designated initializers
       (a C99 extension in C++). Every row is fully designated.
   35  real defects across 15 files, fixed individually.

Three of those 35 were not style:

  - Encyclopedia.cpp discarded a [[nodiscard]] widget id twice. Benign at
    those two sites (the buttons are parented at creation and driven by their
    own onClick), now explicit.
  - GreatPeople.cpp bound a reference to the whole def table and never used
    it. Deleted.
  - AIBuilderController and AISettlerController each store m_difficulty and
    never read it, so DIFFICULTY HAS NO EFFECT on builder or settler
    behaviour. AIMilitaryController and AIResearchPlanner do read theirs.
    Marked [[maybe_unused]] with the gap named in the comment -- wiring it up
    is a behaviour change, not a warning fix.

-Werror=format, -Werror=format-insufficient-args and -Werror=format-security
are now fatal in both the game and the GA targets. A format string that
outnumbers its arguments is undefined behaviour, and an error cannot be
scrolled past.

Behaviour-preserving, verified rather than assumed: sim and worldgen files were
touched, so release 123/123, plus determinism and both goldens green in
release-portable with no re-bless.

+++ src/simulation/diplomacy/WorldCongress.cpp
@@ -199,7 +199,7 @@ Resolution selectResolutionForProposer(const aoc::game::Player& proposer,
     for (uint8_t i = 0; i < order.size(); ++i) { order[i] = i; }
     for (int32_t i = static_cast<int32_t>(order.size()) - 1; i > 0; --i) {
         const int32_t j = rng.nextInt(0, i);
-        std::swap(order[i], order[j]);
+        std::swap(order[static_cast<std::size_t>(i)], order[static_cast<std::size_t>(j)]);
     }
 
     for (uint8_t idx : order) {
